pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meetscribe"
version = "0.1.0"
description = "Modular, extensible, multi-source AI meeting pipeline framework"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    {name = "MeetScribe Team", email = "meetscribe@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Office/Business",
]
keywords = ["meeting", "transcription", "minutes", "ai", "llm", "audio"]

dependencies = [
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
audio = [
    "pydub>=0.25.1",
    "ffmpeg-python>=0.2.0",
]
whisper = [
    "openai>=1.0.0",
]
faster-whisper = [
    "faster-whisper>=0.10.0",
]
gemini = [
    "google-generativeai>=0.3.0",
]
deepgram = [
    "deepgram-sdk>=3.0.0",
]
claude = [
    "anthropic>=0.18.0",
]
google = [
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
]
pdf = [
    "reportlab>=4.0.0",
]
discord = [
    "discord.py>=2.3.0",
    "aiohttp>=3.8.0",
]
webrtc = [
    "aiortc>=1.5.0",
]
obs = [
    "obs-websocket-py>=0.5.0",
]
all = [
    "meetscribe[audio,whisper,faster-whisper,gemini,deepgram,claude,google,pdf,discord]",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
meetscribe = "meetscribe.cli:main"

[project.urls]
Homepage = "https://github.com/m96-chan/MeetScribe"
Documentation = "https://github.com/m96-chan/MeetScribe#readme"
Repository = "https://github.com/m96-chan/MeetScribe"
Issues = "https://github.com/m96-chan/MeetScribe/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["meetscribe*"]

# Black configuration
[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | __pycache__
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["**/venv/**", "**/.venv/**"]

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without explicit strict parameter
    "C901",   # too complex
    "ARG001", # unused function argument
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG", "S101"]

# MyPy configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = [
    "venv",
    ".venv",
    "build",
    "dist",
]

[[tool.mypy.overrides]]
module = [
    "discord.*",
    "aiortc.*",
    "obswebsocket.*",
    "google.*",
    "anthropic.*",
    "openai.*",
    "faster_whisper.*",
    "deepgram.*",
    "reportlab.*",
    "pydub.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
asyncio_mode = "auto"

# Coverage configuration
[tool.coverage.run]
source = ["meetscribe"]
branch = true
omit = [
    "tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "pass",
]
show_missing = true
# Set to 50% as much code requires API keys to test
# Real API-dependent code is tested in integration tests with real credentials
fail_under = 50
